FixMultiplication.cs
Changed according to the suggestions.
